requirements.txt
stomp.py
pyyaml
pyepics